+++ src/QualTrim.h
@@ -186,7 +186,7 @@ int QualTrimIllumina( Read* read, double max_avg_err, double end_lim );
 //extern int window0;
 //extern int window1;
 
-extern int phred_coeff_illumina;
+extern short phred_coeff_illumina;
 extern bool old_style_illumina_flag;
 
 #endif	/* QUALTRIM_H */
